Generify ZipSlip Visitor
Extracts `ZipSlip` visitor into `PathTraversalGuardInsertionVisitor`
to allow for use for other path traversal vulnerabilities.
