+++ src/main/java/org/openrewrite/java/security/PartialPathTraversalVulnerability.java
@@ -22,12 +22,11 @@ import org.openrewrite.analysis.InvocationMatcher;
 import org.openrewrite.analysis.dataflow.DataFlowNode;
 import org.openrewrite.analysis.dataflow.DataFlowSpec;
 import org.openrewrite.analysis.dataflow.Dataflow;
+import org.openrewrite.analysis.search.UsesAllInvocations;
 import org.openrewrite.analysis.trait.expr.Call;
 import org.openrewrite.internal.lang.Nullable;
 import org.openrewrite.java.JavaIsoVisitor;
 import org.openrewrite.java.JavaTemplate;
-import org.openrewrite.java.MethodMatcher;
-import org.openrewrite.java.search.UsesAllMethods;
 import org.openrewrite.java.security.internal.CursorUtil;
 import org.openrewrite.java.tree.Expression;
 import org.openrewrite.java.tree.J;
@@ -40,13 +39,10 @@ import java.util.*;
 import static org.openrewrite.java.security.internal.FileSeparatorUtil.isFileSeparatorExpression;
 
 public class PartialPathTraversalVulnerability extends Recipe {
-    private static final MethodMatcher getCanonicalPathMethodMatcher = new MethodMatcher("java.io.File getCanonicalPath()");
     private static final InvocationMatcher getCanonicalPathMatcher =
-            InvocationMatcher.fromMethodMatcher(getCanonicalPathMethodMatcher);
-
-    private static final MethodMatcher startsWithMethodMatcher = new MethodMatcher("java.lang.String startsWith(java.lang.String)");
+            InvocationMatcher.fromMethodMatcher("java.io.File getCanonicalPath()");
     private static final InvocationMatcher startsWithMatcher =
-            InvocationMatcher.fromMethodMatcher(startsWithMethodMatcher);
+            InvocationMatcher.fromMethodMatcher("java.lang.String startsWith(java.lang.String)");
 
     @Override
     public String getDisplayName() {
@@ -70,10 +66,14 @@ public class PartialPathTraversalVulnerability extends Recipe {
 
     @Override
     public TreeVisitor<?, ExecutionContext> getVisitor() {
-        return Preconditions.check(new UsesAllMethods<>(getCanonicalPathMethodMatcher, startsWithMethodMatcher), new JavaIsoVisitor<ExecutionContext>() {
+        return Preconditions.check(new UsesAllInvocations<>(getCanonicalPathMatcher, startsWithMatcher), new JavaIsoVisitor<ExecutionContext>() {
             @Override
             public J.CompilationUnit visitCompilationUnit(J.CompilationUnit cu, ExecutionContext ctx) {
-                J.CompilationUnit compilationUnit = (J.CompilationUnit) new ZipSlip.ZipSlipComplete<>(false, false).visitNonNull(cu, ctx, getCursor().getParentOrThrow());
+                J.CompilationUnit compilationUnit = (J.CompilationUnit) new PathTraversalGuardInsertionVisitor<>(
+                        ZipSlip.ZIP_ENTRY_GET_NAME,
+                        "doesNotMatter",
+                        false
+                ).visitNonNull(cu, ctx, getCursor().getParentOrThrow());
                 if (cu != compilationUnit) {
                     // The root cause of this vulnerability is Zip Slip, so don't run partial Path
                     return cu;
